pod create: work with uts host namespace
When you have utsns="host" set in containers.conf the podman pod create
command currently fails because it always sets a hostname for the pod.
This is wrong, we only want a hostname with a private uts namespace.
Otherwise we should ignore it.

The test currently cannot work rootless because #13763.

Ref #13714

Signed-off-by: Paul Holzinger <pholzing@redhat.com>