Cirrus: Use fixed netavark/aardvark-dns branch
This is important for the stability of CI in case of a  future backport
that happens to be incompatible with netavark/aardvark `main`.  Since CI
doesn't run very often on the podman `v4.0` branch, an incompatible change
may not be noticed.  Fix this by switching off of the `main` branch onto
a netavark/aardvark release branches.

Signed-off-by: Chris Evich <cevich@redhat.com>