image list: do not reformat `.CreatedAt`
As shown in #14456, the `.CreatedAt` fields for `image list` and of
`image history` can differ by one.  The discussed theory is that the
off-by-one is caused by rounding.

Indeed, the field of `image list` is reformatted.  `image history` is
returning the UNIX time; just as the `.CreatedAt` field should.

I am unable to create a reproducer for the issue but double-checked
what the docker client does: return the UNIX time.

[NO NEW TESTS NEEDED]

Fixes: #14456
Signed-off-by: Valentin Rothberg <vrothberg@redhat.com>