Cirrus: Fix artifact passing when dep. skipped
When passing artifacts from one task to another, if a direct dependency
was not created due to `only_if`, a race condition can occur:  The
descendent task may start running prior to the indirect (but required)
dependency (`build` in this case) completing and uploading artifacts.
Fix this by spamming a `build` dependency into all tasks which require
it (basically everything).

Signed-off-by: Chris Evich <cevich@redhat.com>