Fix unnecessarily installing bats
This tool is pre-installed in all CI VM images, but may be installed by
developers locally.  However, it's not mentioned as a make target
dependency, and the 'already installed' check in the script is broken.
This results in trying to install bats every time CI runs or the make
target is required.  Fix both.

Signed-off-by: Chris Evich <cevich@redhat.com>