implement podman container createspec
first implementation at a command which turns a json specgen into a container.
The spec can be as empty or as complete as users want, as long as an image is provided this command will work.
Also, since this just uses the specgen and create functions, users can pass in all of the run/create flags currently available

a flag that is unique to this command is --start which runs the new container once it is created

resolves #14264

Signed-off-by: Charlie Doern <cdoern@redhat.com>